Added skin management form
Added a form where you can install and unintall skins. Currently, it can only install skins in zip files, but more files may be supported in the future. It's also not possible from within the software to uninstall the default skin, since various parts of KSM Skin Manager depend on it even if it's not being used.